ant seed fix attempt

+++ ACO/Ant.cpp
@@ -43,7 +43,7 @@ Ant::Ant(ACO* aco, ACOGraph* g) {
     }
     // we generate a seed that depends on both the time and the object memory address
     // now each ant will start in a different node at each execution
-    double seed = static_cast<unsigned>(time(0)) + reinterpret_cast<unsigned>(this);
+    int seed = static_cast<int>(time(0)) + reinterpret_cast<int>(this);
     srand(seed);
     int starting_node = rand() % graph->n; // starting node (city)
     solution.push_back(starting_node); // starting node is added to solution for this ant solution
